refactor(tests): lift _make_registry/_load_schema into conftest (Phase A follow-up)
Replaces the deprecated jsonschema.RefResolver pattern used by
test_consumer_external_references_refs.py and
test_consumer_mappings_per_framework_wiring.py. Adds module-scoped
schema_registry fixture so C1 test modules can import _make_registry
without re-duplicating the helper. Migrates the two pre-existing
duplicates in the same commit; net -55 lines of test infrastructure.
